--- src/gameInstance.ts
@@ -1,14 +1,0 @@
-import { EventFunction } from "./event";
-
-export class GameInstance {
-    events: string[];
-    output: string[];
-    queue: EventFunction[];
-    state: any;
-
-    constructor() {
-        this.events = [];
-        this.output = [];
-        this.queue = [];
-    }
-}
